Modified the logic in wp-admin/menu.php and wp-admin/edit-comments.php to change the capability used to determine whether a user can view and modify comments from the back-end from edit_posts to edit_others_posts.
The original code allowed Contributors, who we enable to edit their own
posts, to edit comments by other authors.  This is crazy.  Ideally, a
user with edit_posts should be able to edit their own comments, and a
user with edit_others_posts to edit comments of others, but I'm not up for
the major surgery this would require, especially since we already
provide a means for users to edit their own comments in the front-end.