Modified the theme so that the notification bubble and drop-down list will appear on devices with narrow screens such as tablets and mobile phones.
Added style definitions for notifications for likes of posts and
comments so the special styles invoked for narrow screen devices will
not modify them.

On devices with very narrow screens (< 600 pixels), right-aligned the
notification drop-down list with the bubble and right justified items in
the list to minimise truncation of long items and, if it occurs, show
the most significant part of the notification.